Move functionality for connecting to state and displaying beaches to BeachesContainer
Essentially, I want to be able to CRUD beaches without re-rendering the entire BeachesPage each time.
Ergo, I am moving the entire collection of beaches (and the method of retrieving them) into a separate container.
