+++ src/components/BeachesPage.js
@@ -1,10 +1,8 @@
 import React, { Component } from 'react';
-import Beach from './Beach';
-import { connect } from 'react-redux';
+import BeachesContainer from '../containers/BeachesContainer';
 
 class BeachesPage extends Component {
   render() {
-    const beaches = this.props.beaches; 
     
     return (
       <>
@@ -12,16 +10,10 @@ class BeachesPage extends Component {
           <h1>Your Saved Beaches:</h1>
         </header>
 
-        {Object.keys(beaches).map(
-          beachKey => <Beach key={beachKey} {...beaches[beachKey]} />
-        )}
+        <BeachesContainer />
       </>
     );
   }
 }
 
-const mapStateToProps = state => ({
-  beaches: state.beaches
-});
-
-export default connect(mapStateToProps)(BeachesPage);
+export default BeachesPage;
